pyproject.toml
[project]
name = "electrothon-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.1",
    "google-genai>=1.67.0",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.12",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "pillow>=12.1.1",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.41.0",
    "youtube-transcript-api>=1.2.4",
    "yt-dlp>=2026.3.13",
]
